+++ CHANGES.rst
@@ -4,5 +4,7 @@ Changelog
 0.1-dev (unreleased)
 --------------------
 
+- Add password regeneration option to Reset Password page.
+  [davidjb]
 - Package created using templer
   [JCU eResearch Centre]